requirements.txt
Flask==3.1.0
flask-restful==0.3.10
python-dotenv==1.0.1
mysql-connector-python==8.4.0
